adds mealplans to database
